Rename the button in IG download dialog
Other buttons indicate the file format only and do not do any action. "XLSX" suggests it is doing the same, while it runs the plug-in and downloads the file. I believe it should mention the action that is going to be done. 
How about renaming the button to "Download as XLSX"?